# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open Sky Holidays API",
        "description": "API documentation for Open Sky Holidays",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/services/{slugOrId}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get service details by Slug or ID",
                "description": "Retrieves complete details for a specific travel service using its slug (e.g. visa) or numeric ID.",
                "operationId": "f9abfaa2d656f3b00db640347a9e516e",
                "parameters": [
                    {
                        "name": "slugOrId",
                        "in": "path",
                        "description": "Slug (e.g. \"visa\") or numeric ID of the service",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "visa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service fetched successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            }
        },
...
        "/api/top-header/active": {
            "get": {
                "tags": [
                    "Top Header"
                ],
                "summary": "Get active top header bar data",
                "description": "Retrieves the active top header bar details including email, announcement tagline, button text/link, and social icons.",
                "operationId": "e5bad1fb194783eff7e863e2eb1b4620",
                "responses": {
                    "200": {
                        "description": "Active top header bar retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Active top header bar retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TopHeader"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active top header bar found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No active top header bar found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/top-headers/{id}": {
            "get": {
                "tags": [
                    "Top Header"
                ],
                "summary": "Get a specific top header bar",
                "description": "Retrieves a single top header bar by its ID.",
                "operationId": "9fd1965a0b36693af0a6ed59dfa306c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the top header bar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top header bar retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Top header bar retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TopHeader"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Top header bar not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Top header bar not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/what-we-offers/{whatWeOffer}": {
            "get": {
                "tags": [
                    "What We Offer"
                ],
                "summary": "Get single what we offer item details",
                "description": "Retrieves single what we offer item record by ID.",
                "operationId": "17c521db6af89e406053db812d196604",
                "parameters": [
                    {
                        "name": "whatWeOffer",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "What We Offer item retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "What We Offer item retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/WhatWeOffer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "What We Offer item not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Service": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Visa Assistance Services"
                    },
                    "slug": {
                        "type": "string",
                        "example": "visa-assistance-services"
                    },
                    "about_title": {
                        "type": "string",
                        "example": "Your Gateway to Seamless Global Travel",
                        "nullable": true
                    },
                    "about_description": {
                        "type": "string",
                        "example": "Navigating international visa requirements can be overwhelming.",
                        "nullable": true
                    },
                    "about_image_url": {
                        "description": "About section image URL (JPG, JPEG, PNG, WEBP, AVIF)",
                        "type": "string",
                        "example": "http://127.0.0.1:8000/storage/services/about/visa.jpg",
                        "nullable": true
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "icon": {
                                    "type": "string",
                                    "example": "fa-solid fa-clipboard-list"
                                },
                                "title": {
                                    "type": "string",
                                    "example": "Best Fares"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Competitive pricing on all routes"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "service_items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "process_steps": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "icon": {
                                    "type": "string",
                                    "example": "01"
                                },
                                "title": {
                                    "type": "string",
                                    "example": "Document Review"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "We verify your documents"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "documents": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "why_choose_items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "cta_title": {
                        "type": "string",
                        "example": "Ready To Start Your Journey?",
                        "nullable": true
                    },
                    "cta_description": {
                        "type": "string",
                        "example": "Let us take care of your visa process while you focus on making unforgettable memories.",
                        "nullable": true
                    },
                    "cta_background_image_url": {
                        "description": "CTA banner background image URL (JPG, JPEG, PNG, WEBP, AVIF)",
                        "type": "string",
                        "example": "http://127.0.0.1:8000/storage/services/cta/banner.jpg",
                        "nullable": true
                    },
                    "stats": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "number": {
                                    "type": "string",
                                    "example": "10,000+"
                                },
                                "label": {
                                    "type": "string",
                                    "example": "Visas Processed"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "WhatWeOffer": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Customized Holiday Packages"
                    },
                    "subtitle": {
                        "type": "string",
                        "example": "Tailored for your dream vacation",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "We offer personalized itineraries for individuals, families, and groups.",
                        "nullable": true
                    },
                    "image": {
                        "description": "What we offer image (JPG, JPEG, PNG, WEBP, AVIF)",
                        "type": "string",
                        "example": "what-we-offers/offer1.jpg",
                        "nullable": true
                    },
                    "image_url": {
                        "description": "Full URL to offer image",
                        "type": "string",
                        "example": "http://127.0.0.1:8000/storage/what-we-offers/offer1.jpg",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            },
...
            "TopHeader": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "example": "info@openskyholidays.com"
                    },
                    "tagline": {
                        "type": "string",
                        "example": "The World Is Waiting. One Stop Destination For All Your Tours & Travels Needs."
                    },
                    "button_text": {
                        "type": "string",
                        "example": "Book Your Tour"
                    },
                    "button_url": {
                        "type": "string",
                        "example": "/tours"
                    },
                    "social_links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "platform": {
                                    "type": "string",
                                    "example": "facebook"
                                },
                                "icon": {
                                    "type": "string",
                                    "example": "fa-brands fa-facebook-f"
                                },
                                "url": {
                                    "type": "string",
                                    "example": "https://facebook.com/openskyholidays"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-27T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-27T12:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}